% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{age_dist_graph}
\alias{age_dist_graph}
\title{Make a ggplot object that shows the age distirbution of cases
with indicator for living or dead.}
\usage{
age_dist_graph(data)
}
\arguments{
\item{data}{the nCoV data to run on. Assued to have age categories.}
}
\value{
a ggplot object showing the age distribution of cases
}
\description{
Make a ggplot object that shows the age distirbution of cases
with indicator for living or dead.
}
