% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-descriptions.R
\docType{data}
\name{airport_data}
\alias{airport_data}
\title{Airport data}
\format{
A data frame with 55,113 rows and 12 variables:
\itemize{
  \item \code{ident} airport code
  \item \code{type} type of airport
  \item \code{name} airport name
  \item \code{elevation_ft} airport elevation, in feet
  \item \code{continent} 2 letter code for continent
  \item \code{iso_country} 2 letter code for country
  \item \code{iso_region} 2 letter code for country with 2 number code for region
  \item \code{municipality} municipality in which airport is located
  \item \code{gps_code}
  \item \code{iata_code}
  \item \code{local_code}
  \item \code{coordinates} GPS coordinates
}
}
\usage{
data(airport_data)
}
\description{
A dataset containing information about airports
}
