% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{apply_travel_restrictions}
\alias{apply_travel_restrictions}
\title{Apply a set of travel restrictions to the travel data, reducing or increasing to a proportion of the average travel.}
\usage{
apply_travel_restrictions(travel_data, travel_restrictions_long)
}
\arguments{
\item{travel_data}{Data.frame. Daily travel data that was previously built. We replace the travelers column in this.}

\item{travel_restrictions_long}{Data.frame. Daily travel restrictions, including dates, source location, and proportion of cases.}
}
\description{
Apply a set of travel restrictions to the travel data, reducing or increasing to a proportion of the average travel.
}
