% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{compare_deaths_to_MERS}
\alias{compare_deaths_to_MERS}
\title{Function to automate the comparison of MERS-CoV deaths with
that from the Kudos line list.}
\usage{
compare_deaths_to_MERS(kudos)
}
\arguments{
\item{kudos}{the data from the kudos line list.}
}
\value{
a list with a data frame of results and a ggplot object.
}
\description{
Function to automate the comparison of MERS-CoV deaths with
that from the Kudos line list.
}
