% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{correct_for_Hubei_reporting}
\alias{correct_for_Hubei_reporting}
\title{Function to correct for the changes in reporting in Hubei}
\usage{
correct_for_Hubei_reporting(cum_data, first_date, last_date, tol = 100)
}
\arguments{
\item{first_date}{the first date to infer incidence over}

\item{last_date}{the latest date to infer incidence over}

\item{cumdat}{data frame with the cumulative number of cases}
}
\value{
a version of the inferred incidence data corrected for reportin changes.
}
\description{
Function to correct for the changes in reporting in Hubei
}
