% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_county_imports.R
\name{distrib_county_imports}
\alias{distrib_county_imports}
\title{distrib_county_imports}
\usage{
distrib_county_imports(
  import_sims_clusters,
  airport_attribution,
  local_dir,
  regioncode,
  yr = 2010
)
}
\arguments{
\item{import_sims_clusters}{number of simulations to run}

\item{airport_attribution}{Airport attribution data.frame}

\item{local_dir}{local data directory}

\item{regioncode}{Region/project name}

\item{yr}{Year of county population data}
}
\value{
A data.frame of counties, dates, and nmber of importations
}
\description{
Distribute importations into airports out to counties
}
