% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_to_county_import_distrib.R
\name{do_airport_attribution}
\alias{do_airport_attribution}
\title{do_airport_attribution}
\usage{
do_airport_attribution(
  airports_to_consider,
  airport_cluster_threshold = 80,
  shapefile_path = paste0("data/", regioncode, "/shp/", "counties_2010_", regioncode,
    ".shp"),
  regioncode,
  yr = 2010,
  local_dir = "data/",
  plot = FALSE,
  print_attr_error = FALSE
)
}
\arguments{
\item{airports_to_consider}{data.frame of airports specific to the region of interest}

\item{airport_cluster_threshold}{distance (km) in which airports should be considered clustered.}

\item{shapefile_path}{file path to shapefile for region}

\item{regioncode}{Region/project name}

\item{yr}{Year of county population data}

\item{local_dir}{local data directory}

\item{plot}{logical, whether to plot tesselation maps}
}
\value{
A data.frame of airports and counties with attributions
}
\description{
Cluster Airports in close proximity
}
