% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{epi_curve_reported_cases}
\alias{epi_curve_reported_cases}
\title{Make the epi curve of reported cases. With living and
dead indicated.}
\usage{
epi_curve_reported_cases(data)
}
\arguments{
\item{data}{the nCov data to run on}
}
\value{
a ggplot object showing the epidemiologic curve.
}
\description{
Make the epi curve of reported cases. With living and
dead indicated.
}
