% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{est_daily_incidence}
\alias{est_daily_incidence}
\title{Function to extract approximate epidemic curves
from the cumulative case data.}
\usage{
est_daily_incidence(cum_data, first_date, last_date, na_to_zeros = FALSE)
}
\arguments{
\item{cum_data}{a data frame with cumulative case data in oit}

\item{first_date}{the first date to infer}

\item{second_date}{the last date to infer...shold be iwthin data range}
}
\value{
a data frame with roughly estimated incidence in it
}
\description{
Function to extract approximate epidemic curves
from the cumulative case data.
}
