% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{est_daily_incidence_corrected}
\alias{est_daily_incidence_corrected}
\title{Wrapper function to correct for the changes in reporting in Hubei and merge with data for all incidence}
\usage{
est_daily_incidence_corrected(
  cum_data,
  first_date,
  last_date,
  tol = 100,
  na_to_zeros = FALSE
)
}
\arguments{
\item{cum_data}{data frame with the cumulative number of cases}

\item{first_date}{the first date to infer incidence over}

\item{last_date}{the latest date to infer incidence over}

\item{tol}{error tolerance, in terms of number of cases}
}
\value{
a corrected version of the inferred incidence data corrected for reporting changes in Hubei.
}
\description{
Wrapper function to correct for the changes in reporting in Hubei and merge with data for all incidence
}
