% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_source.R
\name{est_import_detect_dates}
\alias{est_import_detect_dates}
\title{Estimate the detection dates of importations}
\usage{
est_import_detect_dates(sim_res)
}
\arguments{
\item{sim_res}{Vector of importation simulation results}
}
\description{
Estimate the detection dates of importations
}
