% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_travel_data.R
\name{expand_travel_restrict}
\alias{expand_travel_restrict}
\title{Expand the travel restrictions to include every date, to allow for merging with travel data.}
\usage{
expand_travel_restrict(travel_restrictions)
}
\arguments{
\item{travel_restrictions}{data.frame of travel restrictions with columns loc, min, max, p_travel}
}
\description{
Expand the travel restrictions to include every date, to allow for merging with travel data.
}
