% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{find_recent_file}
\alias{find_recent_file}
\title{find_recent_file}
\usage{
find_recent_file(name_start, path, exclude = NULL)
}
\arguments{
\item{name_start}{character string, first letters in file name}

\item{path}{character string, path to folder of interest, end with "/"}

\item{exclude}{character string, patterns to exclude from the file names of interest}
}
\value{
character string, path to most recent file
}
\description{
find_recent_file
}
