% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{fit_ispline}
\alias{fit_ispline}
\title{Function to fit monotonically increasing spline. Returns a function that
can be used to predict the spline for each date.}
\usage{
fit_ispline(dates, obs, df = round(length(obs)/3))
}
\arguments{
\item{dates}{dates of observations}

\item{obs}{the observations}

\item{df}{degrees of freedon.}
}
\value{
a function that takse in some number of dates and gives predictions on those
}
\description{
Function to fit monotonically increasing spline. Returns a function that
can be used to predict the spline for each date.
}
