% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_airport_city}
\alias{get_airport_city}
\title{Get airport city}
\usage{
get_airport_city(airport_code = "ORD")
}
\arguments{
\item{airport_code}{character, airport code}
}
\value{
City of the aiport
}
\description{
These functions are all vectorized
}
