% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_airport_state}
\alias{get_airport_state}
\title{Get airport state}
\usage{
get_airport_state(airport_code = "ORD")
}
\arguments{
\item{airport_code}{character, airport code}
}
\value{
State/province of the airport
}
\description{
Get airport state
}
