% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_attribution.R
\name{get_airports_to_consider}
\alias{get_airports_to_consider}
\title{get_airports_to_consider}
\usage{
get_airports_to_consider(
  airport_monthly_mean_travelers_csv_file,
  states_of_interest,
  travelers_threshold = 10000
)
}
\arguments{
\item{airport_monthly_mean_travelers_csv_file}{Filename of airport monthly mean travelers}

\item{states_of_interest}{States for which to get county populations}

\item{travelers_threshold}{Minimum monthly average travel volume to be included}
}
\value{
A data.frame of clustered airports, dates, and nmber of importations
}
\description{
Query the census API to get the county populations for the states of interest, assigned
 to the given region code.
}
