% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_clean_JHUCSSE_data}
\alias{get_clean_JHUCSSE_data}
\title{Clean crude data from JHUCSSE and aggregated it to state or county level}
\usage{
get_clean_JHUCSSE_data(
  aggr_level = "UID",
  last_date = Sys.time(),
  case_data_dir = "data/case_data",
  save_raw_data = TRUE,
  us_data_only = FALSE
)
}
\arguments{
\item{aggr_level}{"UID", "source", level at which to calculate the cumulative and incident cases. with UID, unassigned locations are dropped}

\item{case_data_dir}{directory where case data is being saved}

\item{save_raw_data}{whether to save raw data locally}

\item{us_data_only}{whether to only pull US data}
}
\value{

}
\description{
Clean crude data from JHUCSSE and aggregated it to state or county level
}
\examples{

}
