% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{get_global_cum}
\alias{get_global_cum}
\title{Get cumulative case counts by Province_state}
\usage{
get_global_cum(df = jhucsse, case_limit = 100)
}
\arguments{
\item{df}{JHUCSSE cleaned data}

\item{case_limit}{limit to Province_State locations above a certain case count}
}
\value{
Cumulative confirmed cases by Province_State.
}
\description{
Get cumulative case counts by Province_state
}
