% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_incidence_data}
\alias{get_incidence_data}
\title{Get incidence data from JHUCSSE}
\usage{
get_incidence_data(
  first_date = ISOdate(2019, 12, 1),
  last_date = Sys.time(),
  pull_github_data = TRUE
)
}
\arguments{
\item{pull_github_data}{}
}
\value{

}
\description{
Get incidence data from JHUCSSE
}
