% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_incidence_fits}
\alias{get_incidence_fits}
\title{Get incidence data from JHUCSSE}
\usage{
get_incidence_fits(
  aggr_level = "UID",
  first_date = ISOdate(2019, 12, 1),
  last_date = Sys.time(),
  case_data_dir = "data/case_data",
  save_raw_data = TRUE,
  us_data_only = FALSE
)
}
\arguments{
\item{aggr_level}{"UID", "source", level at which to calculate the cumulative and incident cases. with UID, unassigned locations are dropped; "source" refers to US state, Chinese province, and national otherwise.}

\item{case_data_dir}{directory where case data is being saved}

\item{save_raw_data}{whether to save raw data locally}

\item{us_data_only}{whether to only pull US data}

\item{check_saved_data}{whether to check locally saved github data}
}
\value{

}
\description{
Get incidence data from JHUCSSE
}
\examples{

}
