% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_JHUCSSE_data}
\alias{get_JHUCSSE_data}
\title{Pull JHU CSSE GitHub data}
\usage{
get_JHUCSSE_data(
  case_data_dir = "data/case_data",
  last_date = Sys.time(),
  save_data = FALSE,
  us_data_only = TRUE,
  append_wiki = TRUE
)
}
\arguments{
\item{case_data_dir}{directory where daily reported case data files are stored by the function.}

\item{last_date}{last date for which to include case data}

\item{save_data}{whether to save raw data locally}

\item{us_data_only}{whether to only pull US data}

\item{append_wiki}{TRUE/FALSE whether to append the data from wikipedia for early china}
}
\value{
NA (saves a CSV of the current data to the data directory)
}
\description{
Pulls the JHUCSSE total case count data up to current date from GitHub.
This version checks what is already saved, and downloads those that are not.
}
