% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{get_mean_imports}
\alias{get_mean_imports}
\title{Get mean importations across geoids}
\usage{
get_mean_imports(
  yr = 2018,
  output_dir = file.path("model_output", "importation"),
  local_dir = "data/",
  n_sim = 100
)
}
\arguments{
\item{yr}{year of population data}

\item{output_dir}{where output files are saved}

\item{local_dir}{directory of the population data}

\item{n_sim}{number of simulations to include}
}
\value{

}
\description{
Get mean importations across geoids
}
