% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_oag_travel_fulldata}
\alias{get_oag_travel_fulldata}
\title{Get OAG travel data}
\usage{
get_oag_travel_fulldata(
  destination = c("CA"),
  destination_type = "state",
  dest_0 = NULL,
  dest_0_type = NULL,
  dest_aggr_level = "city",
  oag_file = "data/complete_OAG_data.csv"
)
}
\arguments{
\item{destination}{destination of interest; can be a vector.}

\item{destination_type}{options: "airport", "city", "state", "country"}

\item{dest_0}{default=NULL; change to specify higher level destination (i.e. dest_0="USA")}

\item{dest_0_type}{default=NULL; must specify if specifying a `dest_0` option.}

\item{dest_aggr_level}{level to which travel will be aggregated for destination. Includes "airport", "city", "state", "country", "metro" (only available for CA currently)}
}
\description{
Get subsetted and cleaned OAG data for a specific destination, using the full OAG data set
}
