% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_airport_clustering.R
\name{imports_airport_clustering}
\alias{imports_airport_clustering}
\title{imports_airport_clustering}
\usage{
imports_airport_clustering(
  imports_sim,
  airport_attribution,
  regioncode,
  local_dir = "data/"
)
}
\arguments{
\item{imports_sim}{single simulation result from importation model}

\item{airport_attribution}{Airport attribution data.frame}

\item{regioncode}{Region/project name}

\item{local_dir}{local data directory}
}
\value{
A data.frame of clustered airports, dates, and nmber of importations
}
\description{
Change importation model parameters so that they match the airport attribution clusters
}
