% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{make_daily_travel}
\alias{make_daily_travel}
\title{Create Daily Travel}
\usage{
make_daily_travel(travel_data, travel_dispersion = 10)
}
\arguments{
\item{travel_data}{monthly travel data}

\item{travel_dispersion}{How dispersed daily travel should be.
-- Set to 10 for very (i.e., most of travel on a couple days)
-- Set to 3  for moderate
-- Set to .01 for none (evenly mixed across days)}
}
\value{
a data frame with randomly distributed travel into days
}
\description{
Function to extract approximate epidemic curves
from the cumulative case data.
}
