% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{make_daily_travel_faster}
\alias{make_daily_travel_faster}
\title{Convert monthly travel to daily travel data -- fast
- When we have already built the daily data, we reuse that and just fill in the new daily volume each time}
\usage{
make_daily_travel_faster(
  travel_data,
  travel_data_daily,
  travel_dispersion = 10
)
}
\arguments{
\item{travel_data}{Data.frame. Monthly travel data with columns travelers, t_month, and days_month}

\item{travel_data_daily}{Data.frame. Daily travel data that was previously built. We replace the travelers column in this.}

\item{travel_dispersion}{Numeric. Value defining how evenly distributed daily travel is across a month.}
}
\description{
Convert monthly travel to daily travel data -- fast
- When we have already built the daily data, we reuse that and just fill in the new daily volume each time
}
