% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{make_input_data}
\alias{make_input_data}
\title{Make input data}
\usage{
make_input_data(
  incid_data,
  travel_data,
  pop_data,
  dest_aggr_level,
  shift_incid_days = NA
)
}
\arguments{
\item{shift_incid_days}{}
}
\value{

}
\description{
Produce combined input data.
}
