% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{make_meanD}
\alias{make_meanD}
\title{Make MeanD Matrix}
\usage{
make_meanD(
  input_data,
  n_sim,
  incub_mean_log,
  incub_sd_log,
  inf_period_hosp_mean_log,
  inf_period_hosp_sd_log,
  inf_period_nohosp_mean,
  inf_period_nohosp_sd
)
}
\arguments{
\item{inf_period_nohosp_sd}{}
}
\value{

}
\description{
Make MeanD Matrix
}
