% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{OR_table_age}
\alias{OR_table_age}
\title{Make a table capturing the odds ratio of death by age cat}
\usage{
OR_table_age(data, combine_CIs = TRUE)
}
\arguments{
\item{data}{the data inclding an age_cat column}

\item{combine_CIs}{should we combine the confidence intervals
into a single entry, or hav two separate columns}
}
\value{
a data frame with includein colums with OR and CI
}
\description{
Make a table capturing the odds ratio of death by age cat
}
