% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{OR_table_gender}
\alias{OR_table_gender}
\title{Make a table capturing the odds ratio of death by geneder}
\usage{
OR_table_gender(data)
}
\arguments{
\item{data}{}
}
\value{
a data frame with includein colums with OR and CI
}
\description{
Make a table capturing the odds ratio of death by geneder
}
