% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUtils.R
\name{plot_imports_stackeddest}
\alias{plot_imports_stackeddest}
\title{Barchart plot of importations stacked by source}
\usage{
plot_imports_stackeddest(
  data = imports_mean \%>\% rename(amount = import_mean),
  t_limits = lubridate::as_date(c("2020-01-01", "2020-04-20")),
  axes_text_size = c(9, 9),
  leg_text_size = 8,
  leg_title_size = 10,
  time_int = 5,
  n_cols = 20
)
}
\arguments{
\item{data}{data of importations to plot. Must have a columns c("t", "loc", "amount")}

\item{t_limits}{Date limits of plot}

\item{axes_text_size}{axis value font size (2 values required)}

\item{time_int}{time interval for x-axis}

\item{n_cols}{number of colors/locations to keep individually}
}
\value{
a corrected version of the inferred incidence data corrected for reporting changes in Hubei.
}
\description{
Barchart plot of importations stacked by source
}
