% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicEpiAnalyses.R
\name{plot_incidence_ests_report}
\alias{plot_incidence_ests_report}
\title{Function to plot the estimated and reported case counts}
\usage{
plot_incidence_ests_report(
  conf_cases = jhucsse,
  incid_ests = incidence_data,
  locations = "All",
  ncol_facet = 2
)
}
\arguments{
\item{conf_cases}{Confirmed case data from JHU CSSE}

\item{incid_ests}{Estimated incidence}

\item{locations}{Locations to plot, can be a vector}
}
\value{
a corrected version of the inferred incidence data corrected for reporting changes in Hubei.
}
\description{
Function to plot the estimated and reported case counts
}
