% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUtils.R
\name{plot_rep_cases}
\alias{plot_rep_cases}
\title{Facet plot of cumulative and incident cases at locations desired}
\usage{
plot_rep_cases(
  case_data = jhucsse,
  state = "NY",
  county_name = c("New York", "Nassau"),
  aggr_locs = FALSE,
  date_limits = c("2020-01-15", "2020-03-15"),
  ncol_facet = 2
)
}
\arguments{
\item{case_data}{Confirmed case data from JHU CSSE}

\item{state}{State abbreviation; NULL if not a state}

\item{county_name}{Vector of counties of interest; NULL if not wanting counties}

\item{aggr_locs}{whether aggregated cases across specified states or counties is wanted}

\item{date_limits}{date boundaries desired for the plot}

\item{ncol_facet}{number of columns in the facet plotting}
}
\value{
a facet plot of confirmed and incident cases
}
\description{
Facet plot of cumulative and incident cases at locations desired
}
