% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-descriptions.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{Population data}
\format{A data frame with 5 variables:
\itemize{
  \item \code{country_name} country name
  \item \code{country} country code
  \item \code{pop} population
  \item \code{source} if the source is the country, then the country code; if the source is sub-country level, then it is a code for the source
  \item \code{state_name} full name of US states
}}
\usage{
data(pop_data)
}
\description{
A dataset containing populations for each source
}
