% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadUtils.r
\name{read_JHUCSSE_cases}
\alias{read_JHUCSSE_cases}
\title{Reads in the JHUCSSE total case count data up
until (and including) a given dat.}
\usage{
read_JHUCSSE_cases(last_time, append_wiki, print_file_path = FALSE)
}
\arguments{
\item{last_time}{the last time to consider data from}

\item{append_wiki}{sjpi;d we also append data from wikipedia.}

\item{print_file_path}{whether or not to print the file path}
}
\value{
a data frame with the basic data.
}
\description{
Reads in the JHUCSSE total case count data up
until (and including) a given dat.
}
