% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_source.R
\name{run_daily_import_model}
\alias{run_daily_import_model}
\title{Run the full simulation of daily importations}
\usage{
run_daily_import_model(
  input_data,
  travel_data_monthly,
  travel_data_daily,
  travel_dispersion = 3,
  travel_restrictions = NULL,
  allow_travel_variance = FALSE,
  tr_inf_redux = 0,
  get_detection_time = FALSE,
  param_list = list(incub_mean_log = log(5.89), incub_sd_log = log(1.74),
    inf_period_nohosp_mean = 15, inf_period_nohosp_sd = 5, inf_period_hosp_mean_log =
    1.23, inf_period_hosp_sd_log = 0.79)
)
}
\arguments{
\item{input_data}{full importation input data, including case, travel, and population data}

\item{travel_data_monthly}{monthly travel data between sources and destinations}

\item{travel_dispersion}{how evenly the monthly travel should be distributed across days}

\item{travel_restrictions}{data.frame of travel restrictions}

\item{allow_travel_variance}{whether to sample from the travel variance}

\item{tr_inf_redux}{proportion reduction in travel when individuals are infected}

\item{get_detection_time}{logical; return importation detection or not}

\item{n_sim}{number of simulations to run}

\item{meanD_mat}{matrix of mean duration during which infected individuals can travel}

\item{u_origin}{reporting rate, origin}

\item{time_inftodetect}{Time from infection to detection}

\item{incub_mean_log}{log mean incubation period}

\item{incub_sd_log}{log sd of incubation period}

\item{inf_period_hosp_mean_log}{infectious period of hospitalized, log-mean}

\item{inf_period_hosp_sd_log}{infectious period of hospitalized, log-sd}

\item{inf_period_nohosp_mean}{infectious period of non-hospitalized, shape}

\item{inf_period_nohosp_sd}{infectious period of non-hospitalized, scale}

\item{project_name}{project name, if saving in the function}

\item{batch}{run batch, if saving in the function}

\item{version}{run version, if saving in the function}

\item{print_progress}{logical, whether to print the progress of the simulations}

\item{cores}{number of cores for parallel processing}

\item{time_inftotravel}{time from infection to traveling}
}
\value{
list consisting of two objects: 1) an array of importations by date, location, and simulation, 2) a dataframe with negative binomial parameters for each location and date
}
\description{
Run the full simulation of daily importations
}
