% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_to_county_import_distrib.R
\name{run_full_distrib_imports}
\alias{run_full_distrib_imports}
\title{run_full_distrib_imports}
\usage{
run_full_distrib_imports(
  states_of_interest = c("CA", "NV", "WA", "OR", "AZ"),
  regioncode = "west-coast-AZ-NV",
  yr = 2010,
  travelers_threshold = 10000,
  airport_cluster_threshold = 80,
  mean_travel_file = file.path("data", "travel_mean.csv"),
  shapefile_path = NULL,
  model_output_dir = file.path("model_output", "importation"),
  local_dir = "data",
  plot = FALSE,
  cores = 5,
  n_sim = 10
)
}
\arguments{
\item{states_of_interest}{States for which to get county populations}

\item{regioncode}{Region/project name}

\item{yr}{Year of county population data}

\item{travelers_threshold}{Minimum monthly average travel volume to be included}

\item{airport_cluster_threshold}{Distance by which airports can be separated to be in the same cluster, in km; Haversine distance.}

\item{mean_travel_file}{Filename of monthly mean travelers into each airport in the region}

\item{local_dir}{local data directory}

\item{imports_sim_file}{file name root for each importation simulation.}
}
\value{
A data.frame of clustered airports, dates, and nmber of importations
}
\description{
Sum importation counts to airport attribution clusters
Airport clusters are generated using `do_airport_attribution` for aiports in close proximity to 
each other e.g. ORD/MDY, DCA/IAD, SFO/OAK
}
