% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_source.R
\name{run_importations}
\alias{run_importations}
\title{Run importation sims}
\usage{
run_importations(
  n_sim = 100,
  cores = 5,
  get_detection_time = FALSE,
  travel_dispersion = 3,
  allow_travel_variance = FALSE,
  print_progress = TRUE,
  output_dir = file.path("output", paste0(paste(dest, collapse = "+"), "_",
    as.Date(Sys.Date()))),
  param_list = list(incub_mean_log = log(5.89), incub_sd_log = log(1.74),
    inf_period_nohosp_mean = 15, inf_period_nohosp_sd = 5, inf_period_hosp_mean_log =
    1.23, inf_period_hosp_sd_log = 0.79, p_report_source = c(0.05, 0.25))
)
}
\arguments{
\item{n_sim}{numeric, number of simulations}

\item{cores}{numeric, number of cores to run in parallel}

\item{get_detection_time}{logical}

\item{travel_dispersion}{numeric}

\item{allow_travel_variance}{logical}

\item{print_progress}{logical}

\item{param_list}{list, with the following elements
\itemize{
  \item \code{incub_mean_log} numeric, the mean_log parameter for the lnorm distribution for the incubation period
  \item \code{incub_sd_log} numeric, the sd_log parameter for the lnorm distribution for the incubation period
  \item \code{inf_period_nohosp_mean} numeric, the mean parameter of a truncated normal distribution for the infectious period for non-hospitalized infections
  \item \code{inf_period_nohosp_sd} numeric, the sd parameter of a truncated normal distribution for the infectious period for non-hospitalized infections
  \item \code{inf_period_hosp_mean_log} numeric, the log-mean parameter of a log-normal distribution for the infectious period for time to hospitalization
  \item \code{inf_period_hosp_sd_log} numeric, the log-sd parameter of a log-normal distribution for the infectious period for time to hospitalization
  \item \code{p_report_source} numeric vector of length 2, currently the probability of reporting by source with the first indicating Hubei reporting and second indicating everywhere else (UPDATE WITH THESE https://cmmid.github.io/topics/covid19/severity/global_cfr_estimates.html)
  \item \code{shift_incid_days} lag from infection to report
  \item \code{delta} days per time period
}}

\item{dest}{character string, name of destination to simulate importations for}

\item{dest_type}{character string, options: "airport", "city", "state", "country"}

\item{dest_0}{optional character string, specify a higher level destination (i.e. dest_0="USA"), default NULL}

\item{dest_0_type}{optional character string, must specify if specifying a `dest_0` option; default=NULL}

\item{dest_aggr_level}{character string, level to which travel will be aggregated for destination. Includes "airport", "city", "state", "country", "metro" (only available for CA currently)}

\item{project_name}{character string}

\item{version}{character string}

\item{batch}{character string}

\item{end_date}{Date, last import date}

\item{n_top_dests}{Number of destinations to include, ranked by volume; default (Inf) is all.}
}
\value{

}
\description{
Run importation sims
}
