% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_to_county_import_distrib.R
\name{setup_airport_attribution}
\alias{setup_airport_attribution}
\title{setup_airport_attribution}
\usage{
setup_airport_attribution(
  states_of_interest,
  regioncode,
  yr,
  local_dir = "data/",
  write_county_shapefiles,
  mean_travel_file = file.path("data", "travel_mean.csv"),
  travelers_threshold = 10000,
  airport_cluster_threshold = 80,
  plot = FALSE,
  cores = 4
)
}
\arguments{
\item{states_of_interest}{States for which to get county populations}

\item{regioncode}{Region/project name}

\item{yr}{Year of county population data}

\item{local_dir}{local data directory}

\item{write_county_shapefiles}{logical}

\item{mean_travel_file}{Filename of monthly mean travelers into each airport in the region}

\item{airport_cluster_threshold}{distance (km) in which airports should be considered clustered.}

\item{plot}{logical, whether to plot tesselation maps}

\item{cores}{number of cores to run in parallel}

\item{airports_to_consider}{data.frame of airports specific to the region of interest}
}
\value{
A data.frame of airports and counties with attributions
}
\description{
Cluster Airports in close proximity
}
