% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_source.R
\name{setup_and_run_importations}
\alias{setup_and_run_importations}
\title{Set up and run importation sims}
\usage{
setup_and_run_importations(
  dest = "UT",
  dest_type = c("state"),
  dest_country = "USA",
  dest_aggr_level = c("airport"),
  project_name = "Utah_import",
  version = "global",
  batch = "1st",
  first_date = ISOdate(2019, 12, 1),
  last_date = Sys.time(),
  update_case_data = TRUE,
  case_data_dir = "data/case_data",
  check_saved_data = TRUE,
  save_data = TRUE,
  get_travel = TRUE,
  n_sim = 100,
  cores = 4,
  get_detection_time = FALSE,
  travel_dispersion = 3,
  allow_travel_variance = FALSE,
  print_progress = TRUE,
  get_nbinom_params = TRUE,
  param_list = list(incub_mean_log = log(5.89), incub_sd_log = log(1.74),
    inf_period_nohosp_mean = 15, inf_period_nohosp_sd = 5, inf_period_hosp_shape = 0.75,
    inf_period_hosp_scale = 5.367, p_report_source = c(0.05, 0.25), shift_incid_days =
    -10, delta = 1)
)
}
\arguments{
\item{dest}{character string, name of destination to simulate importations for}

\item{dest_type}{character string, options: "airport", "city", "state", "country"}

\item{dest_aggr_level}{character string, level to which travel will be aggregated for destination. Includes "airport", "city", "state", "country", "metro" (only available for CA currently)}

\item{project_name}{character string}

\item{version}{character string}

\item{batch}{character string}

\item{n_sim}{numeric, number of simulations}

\item{cores}{numeric, number of cores to run in parallel}

\item{get_detection_time}{logical}

\item{travel_dispersion}{numeric}

\item{allow_travel_variance}{logical}

\item{print_progress}{logical}

\item{param_list}{list, with the following elements
\itemize{
  \item \code{incub_mean_log} numeric, the mean_log parameter for the lnorm distribution for the incubation period
  \item \code{incub_sd_log} numeric, the sd_log parameter for the lnorm distribution for the incubation period
  \item \code{inf_period_nohosp_mean} numeric, the mean parameter of a truncated normal distribution for the infectious period for non-hospitalized infections
  \item \code{inf_period_nohosp_sd} numeric, the sd parameter of a truncated normal distribution for the infectious period for non-hospitalized infections
  \item \code{inf_period_hosp_shape} numeric, the shape parameter of a Gamma distribution for the infectious period for time to hospitalization
  \item \code{inf_period_hosp_scale} numeric, the scale parameter of a Gamma distribution for the infectious period for time to hospitalization
  \item \code{p_report_source} numeric vector of length 2, currently the probability of reporting by source with the first indicating Hubei reporting and second indicating everywhere else (UPDATE WITH THESE https://cmmid.github.io/topics/covid19/severity/global_cfr_estimates.html)
  \item \code{shift_incid_days} lag from infection to report
  \item \code{delta} days per time period
}}

\item{dest_0}{optional character string, specify a higher level destination (i.e. dest_0="USA"), default NULL}

\item{dest_0_type}{optional character string, must specify if specifying a `dest_0` option; default=NULL}

\item{end_date}{Date, last import date}
}
\value{

}
\description{
Set up and run importation sims
}
