% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{update_JHUCSSE_github_data}
\alias{update_JHUCSSE_github_data}
\title{Pull JHU CSSE GitHub data}
\usage{
update_JHUCSSE_github_data(
  case_data_dir = "data/case_data",
  last_date = Sys.Date(),
  check_saved_data = FALSE,
  save_data = FALSE
)
}
\arguments{
\item{case_data_dir}{directory where daily reported case data files are stored by the function.}

\item{last_date}{last date for which to include case data}

\item{check_saved_data}{whether to check for existing saved case data}

\item{save_data}{whether to save the cleaned and combined data}
}
\value{
NA (saves a CSV of the current data to the data directory)
}
\description{
Pulls the JHUCSSE total case count data up to current date from GitHub.
This version checks what is already saved, and downloads those that are not.
}
