% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-descriptions.R
\docType{data}
\name{airport_attribution}
\alias{airport_attribution}
\title{Airport attribution data for China}
\format{A data frame with 507 rows and 3 variables:
\itemize{
  \item \code{Province} the name of a Chinese province
  \item \code{airport_iata} airport code
  \item \code{attribution} the percentage of people from that province that use that airport, such that the sum of the attributions for each province equals 1
}}
\usage{
data(airport_attribution)
}
\description{
A dataset containing information about which Chinese airports the people in
different Chinese provinces use.
}
