% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_source.R
\name{calc_nb_import_pars}
\alias{calc_nb_import_pars}
\title{Get negative binomial estimates for each time and destination}
\usage{
calc_nb_import_pars(importation_sim)
}
\arguments{
\item{importation_sim}{4D array outputted from the importation model}

\item{version}{}
}
\description{
Get negative binomial estimates for each time and destination
}
