% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_airport_city}
\alias{get_airport_city}
\title{Get Airport Location Data
These functions are all vectorized}
\usage{
get_airport_city(airport_code = "ORD")
}
\description{
Get Airport Location Data
These functions are all vectorized
}
