% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_airport_country}
\alias{get_airport_country}
\title{Get airport country}
\usage{
get_airport_country(airport_code = "ORD")
}
\arguments{
\item{airport_code}{character, airport code}
}
\value{
ISO3 code for the country where the airport is
}
\description{
Get airport country
}
\examples{
get_airport_country()

}
