% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_to_county_import_distrib.R
\name{get_county_pops}
\alias{get_county_pops}
\title{get_county_pops}
\usage{
get_county_pops(
  states_of_interest,
  regioncode,
  yr = 2010,
  local_dir = "data/",
  write_county_shapefiles = TRUE
)
}
\arguments{
\item{states_of_interest}{States for which to get county populations}

\item{regioncode}{Region/project name}

\item{yr}{Year of county population data}

\item{local_dir}{local data directory}
}
\value{
A data.frame of clustered airports, dates, and nmber of importations
}
\description{
These are all of the functions needed to do the airport-to-county importation distribution process.
}
\details{
Query the census API to get the county populations for the states of interest, assigned
 to the given region code.
 
Assumes census_api_key is already called by the calling client with 
 `census_api_key(key="c235e1b5620232fab506af060c5f8580604d89c1", install=TRUE)``
     and
 `options(tigris_use_cache = TRUE)`
}
