% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{get_incidence_data}
\alias{get_incidence_data}
\title{Get incidence data from JHUCSSE}
\usage{
get_incidence_data(
  first_date = ISOdate(2019, 12, 1),
  last_date = Sys.time(),
  update_case_data = TRUE,
  case_data_dir = "data/case_data",
  check_saved_data = TRUE,
  save_data = TRUE
)
}
\arguments{
\item{update_case_data}{whether to update the case data from the JHUCSSE github or just use data that are part of this package}

\item{case_data_dir}{directory where case data is being saved}

\item{check_saved_data}{whether to check locally saved github data}

\item{save_data}{whether to save data locally}
}
\value{

}
\description{
Get incidence data from JHUCSSE
}
\examples{

}
