% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_to_county_import_distrib.R
\name{imports_airport_clustering}
\alias{imports_airport_clustering}
\title{imports_airport_clustering}
\usage{
imports_airport_clustering(
  imports_sim,
  airport_attribution,
  model_output_dir = "model_output/importation"
)
}
\arguments{
\item{imports_sim}{single simulation result from importation model}

\item{airport_attribution}{Airport attribution data.frame}

\item{regioncode}{Region/project name}

\item{local_dir}{local data directory}
}
\value{
A data.frame of clustered airports, dates, and nmber of importations
}
\description{
Sum importation counts to airport attribution clusters
Airport clusters are generated using `do_airport_attribution` for aiports in close proximity to 
each other e.g. ORD/MDY, DCA/IAD, SFO/OAK
}
