% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUtils.R, R/plot_importations.R
\name{plot_imports_stackeddest}
\alias{plot_imports_stackeddest}
\title{Barchart plot of importations stacked by source}
\usage{
plot_imports_stackeddest(
  data = imports_mean,
  t_limits = lubridate::as_date(c("2020-01-01", "2020-04-20")),
  y.size = 9,
  x.size = 9,
  leg.text.size = 8,
  leg.title.size = 10,
  time_int = 5,
  n_cols = 20
)

plot_imports_stackeddest(
  data = imports_mean,
  t_limits = lubridate::as_date(c("2020-01-01", "2020-04-20")),
  y.size = 9,
  x.size = 9,
  leg.text.size = 8,
  leg.title.size = 10,
  time_int = 5,
  n_cols = 20
)
}
\arguments{
\item{data}{data of importations to plot. Must have a columns c("t", "loc", "amount")}

\item{t_limits}{Date limits of plot}

\item{time_int}{time interval for x-axis}

\item{n_cols}{number of colors/locations to keep individually}

\item{axes_text_size}{axis value font size (2 values required)}

\item{incid_ests}{Estimated incidence}

\item{locations}{Locations to plot, can be a vector}

\item{ncol_facet}{number of columns in the facet plotting}
}
\value{
a corrected version of the inferred incidence data corrected for reporting changes in Hubei.

a corrected version of the inferred incidence data corrected for reporting changes in Hubei.
}
\description{
Barchart plot of importations stacked by source

Barchart plot of importations stacked by source
}
