% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{pull_JHUCSSE_github_data}
\alias{pull_JHUCSSE_github_data}
\title{Pull JHU CSSE GitHub data}
\usage{
pull_JHUCSSE_github_data()
}
\value{
NA (saves a CSV of the current data to the data directory)
}
\description{
Pulls the JHUCSSE total case count data up to current date from GitHub.
This version checks what is already saved, and downloads those that are not.
Eventually, we would like automate this.
}
