% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_model_source.R
\name{run_daily_import_model_par}
\alias{run_daily_import_model_par}
\title{Run the full simulation of daily importations}
\usage{
run_daily_import_model_par(
  n_sim = 10000,
  input_data,
  travel_data_monthly = travel_data_monthly,
  travel_dispersion = 3,
  travel_restrictions = data.frame(loc = "Hubei", min = "2020-01-25", max =
    "2020-04-01", p_travel = 0),
  allow_travel_variance = FALSE,
  meanD_mat,
  tr_inf_redux,
  u_origin,
  get_detection_time = FALSE,
  time_inftotravel,
  time_inftodetect,
  incub_mean_log,
  incub_sd_log,
  inf_period_hosp_shape,
  inf_period_hosp_scale,
  inf_period_nohosp_mean,
  inf_period_nohosp_sd,
  project_name,
  batch,
  version,
  print_progress = TRUE,
  cores = 4
)
}
\arguments{
\item{n_sim}{number of simulations to run}

\item{input_data}{full importation input data, including case, travel, and population data}

\item{travel_data_monthly}{monthly travel data between sources and destinations}

\item{travel_dispersion}{how evenly the monthly travel should be distributed across days}

\item{travel_restrictions}{data.frame of travel restrictions}

\item{allow_travel_variance}{whether to sample from the travel variance}

\item{meanD_mat}{matrix of mean duration during which infected individuals can travel}
}
\description{
Run the full simulation of daily importations
}
